Made flags into an int

+++ MelonCameraMod/CameraMod.cs
@@ -139,7 +139,7 @@ namespace MelonCameraMod
                 camera.orthographic = config.Orthographic;
                 camera.orthographicSize = config.OrthographicSize;
                 camera.rect = config.Rect;
-                camera.cullingMask = config.CullingMask;
+                camera.cullingMask = (int)config.CullingMask;
 
                 camera.eventMask = 0;
                 camera.stereoTargetEye = StereoTargetEyeMask.None;
